+++ kinto/core/permission/postgresql/__init__.py
@@ -203,6 +203,9 @@ class Permission(PermissionBase):
     def get_accessible_objects(self, principals, bound_permissions=None):
         principals_values = ','.join(["('%s')" % p for p in principals])
         if bound_permissions is None:
+            # Return all objects on which the specified principals have some
+            # permissions.
+            # (e.g. permissions endpoint which lists everything)
             query = """
             WITH user_principals AS (
               VALUES %(principals)s
@@ -212,6 +215,11 @@ class Permission(PermissionBase):
               JOIN user_principals
                 ON (principal = user_principals.column1);
             """ % dict(principals=principals_values)
+        elif len(bound_permissions) == 0:
+            # If the list of object permissions to filter on is empty, then
+            # do not bother querying the backend. The result will be empty.
+            # (e.g. root object /buckets)
+            return {}
         else:
             perms = [(o.replace('*', '.*'), p) for (o, p) in bound_permissions]
             perms_values = ','.join(["('%s', '%s')" % p for p in perms])
@@ -355,11 +363,20 @@ class Permission(PermissionBase):
         if len(object_id_list) == 0:
             return
 
+        object_ids_values = ','.join(["('^%s$')" % o.replace('*', '.*')
+                                      for o in object_id_list])
         query = """
+        WITH object_ids AS (
+          VALUES %(object_ids_values)s
+        )
         DELETE FROM access_control_entries
-         WHERE object_id IN :object_id_list;"""
+         USING object_ids
+         WHERE object_id ~ column1;"""
+        safeholders = {
+            'object_ids_values': object_ids_values
+        }
         with self.client.connect() as conn:
-            conn.execute(query, dict(object_id_list=tuple(object_id_list)))
+            conn.execute(query % safeholders)
 
 
 def load_from_config(config):